Cleaned Up.
